pyproject.toml
[project]
name = "pd-agents"
version = "0.1.0"
description = "Phase 1: Iterated Prisoner's Dilemma benchmark for AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "streamlit>=1.36.0",
]

[project.scripts]
pdbench = "pdbench.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
